requirements.txt
fastmcp>=2.10.5
aiohttp>=3.9.0
redis>=5.0.0
pandas>=2.0.0
numpy>=1.24.0
google-cloud-bigquery>=3.11.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tenacity>=8.2.0
prometheus-client>=0.17.0
structlog>=23.1.0
